requirements.txt
numpy~=2.1.1
opencv-python~=4.10.0.84
pillow~=10.4.0
scikit-image~=0.24.0
scikit-learn~=1.5.2
scipy~=1.14.1